requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
duckdb==1.2.0
Faker==36.1.1
fastapi==0.115.8
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
numpy==2.2.3
pandas==2.2.3
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uv==0.6.0
uvicorn==0.34.0
